Переработка БД, DTO и мапперов

+++ src/main/java/org/study/grabli_application/repository/StreetObjectTypeRepository.java
@@ -3,5 +3,5 @@ package org.study.grabli_application.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.study.grabli_application.entity.StreetObjectType;
 
-public interface StreetObjectTypeRepository extends JpaRepository<StreetObjectType, Long> {
+public interface StreetObjectTypeRepository extends JpaRepository<StreetObjectType, Integer> {
 }
